oya-back/requirements.txt
fastapi==0.70.0
pydantic==1.8.2
python_dateutil==2.8.2
SQLAlchemy==1.4.27
psycopg2==2.9.2
python-decouple==3.5
uvicorn==0.15.0
alembic==1.7.5

